+++ src/main.rs
@@ -1,8 +1,12 @@
 pub mod cli;
 pub mod crypto;
 pub mod types;
+pub mod utils;
 
 fn main() {
+    // Derive keys from mnemonic:
     // cli::hdkeys::derive();
-    cli::import::import();
+
+    // Export secret key from wallet.toml
+    cli::export::export();
 }
